refactor: extract BasePreprocessor with Template Method pattern
Move shared preprocessor logic (block parsing, Liquid rendering,
index lookup) into BasePreprocessor. LutamlPreprocessor now only
implements format-specific hooks: load_lutaml_file dispatches to
Express::Parsers::Exp or Uml::Parsers::Dsl, update_repo handles
ExpFile/Repository unwrapping and remark decoration, template uses
the custom Liquid environment with keyiterator tag and filters.
