+++ lib/metanorma/plugin/lutaml/lutaml_preprocessor.rb
@@ -1,21 +1,15 @@
 # frozen_string_literal: true
 
-require "liquid"
-require "asciidoctor"
-require "asciidoctor/reader"
-require "lutaml"
-require "metanorma/plugin/lutaml/utils"
-require "metanorma/plugin/lutaml/asciidoctor/preprocessor"
 require "metanorma/plugin/lutaml/express_remarks_decorator"
 
 module Metanorma
   module Plugin
     module Lutaml
-      # Class for processing Lutaml files
-      class LutamlPreprocessor < ::Asciidoctor::Extensions::Preprocessor
-        include Utils
-
-        REMARKS_ATTRIBUTE = "remarks"
+      # Preprocessor for EXPRESS schema formats (lutaml, lutaml_express,
+      # lutaml_express_liquid). Parses EXPRESS files via the lutaml/expressir
+      # gems, decorates remarks with relative path resolution, and renders
+      # Liquid templates with the EXPRESS-specific Liquid environment.
+      class LutamlPreprocessor < BasePreprocessor
         EXPRESS_PREPROCESSOR_REGEX = %r{
           ^                            # Start of line
           \[                           # Opening bracket
@@ -30,117 +24,38 @@ module Metanorma
           \]                           # Closing bracket
         }x
 
-        def process(document, reader) # rubocop:disable Metrics/MethodLength
-          r = Asciidoctor::PreprocessorNoIfdefsReader.new(document,
-                                                          reader.lines)
-          input_lines = r.readlines.to_enum
-
-          express_indexes = Utils.parse_document_express_indexes(
-            document,
-            input_lines,
-          )
-
-          result_content = process_input_lines(
-            document: document,
-            input_lines: input_lines,
-            express_indexes: express_indexes,
-          )
-
-          Asciidoctor::PreprocessorNoIfdefsReader.new(document, result_content)
-        end
-
         protected
 
         def lutaml_liquid?(line)
           line.match(EXPRESS_PREPROCESSOR_REGEX)
         end
 
-        def load_express_lutaml_file(document, file_path)
-          ::Lutaml::Parser.parse(
-            File.new(
-              Utils.relative_file_path(document, file_path),
-              encoding: "UTF-8",
-            ),
-          )
-        end
+        def load_lutaml_file(document, file_path, _options)
+          full_path = Utils.relative_file_path(document, file_path)
 
-        private
-
-        def process_input_lines(document:, input_lines:, express_indexes:)
-          result = []
-          loop do
-            result.push(
-              *process_text_blocks(document, input_lines, express_indexes),
-            )
+          file = File.new(full_path, encoding: "UTF-8")
+          if full_path.end_with?(".exp")
+            ::Lutaml::Express::Parsers::Exp.parse(file)
+          else
+            ::Lutaml::Uml::Parsers::Dsl.parse(file)
           end
-          result
         end
 
-        def process_text_blocks(document, input_lines, express_indexes) # rubocop:disable Metrics/AbcSize,Metrics/CyclomaticComplexity,Metrics/MethodLength
-          line = input_lines.next
-          block_header_match = lutaml_liquid?(line)
-
-          return [line] if block_header_match.nil?
-
-          index_names = block_header_match[:index_names].split(";").map(&:strip)
-          context_name = block_header_match[:context_name].strip
-
-          options = (block_header_match[:options] &&
-            parse_options(block_header_match[:options].to_s.strip)) || {}
-
-          end_mark = input_lines.next
-
-          render_liquid_template(
-            document: document,
-            lines: extract_block_lines(input_lines, end_mark),
-            index_names: index_names,
-            context_name: context_name,
-            options: options,
-            indexes: express_indexes,
-          )
+        def index_type_name
+          "EXPRESS"
         end
 
-        def extract_block_lines(input_lines, end_mark)
-          block = []
-          while (block_line = input_lines.next) != end_mark
-            block.push(block_line)
-          end
-          block
-        end
-
-        def gather_context_liquid_items( # rubocop:disable Metrics/AbcSize,Metrics/MethodLength,Metrics/ParameterLists
-          index_names:, document:, indexes:, options: {}
-        )
-          index_names.map do |path| # rubocop:disable Metrics/BlockLength
-            if indexes[path] && indexes[path][:model]
-              repo = indexes[path][:model]
-              repo = update_repo(options, repo)
-              indexes[path][:liquid_drop] ||= repo.to_liquid
-            else
-              full_path = Utils.relative_file_path(document, path)
-              unless File.file?(full_path)
-                raise StandardError.new(
-                  "Unable to load EXPRESS index for `#{path}`, " \
-                  "please define it at `:lutaml-express-index:` or specify " \
-                  "the full path.",
-                )
-              end
-              repo = load_express_lutaml_file(document, path)
-              repo = update_repo(options, repo)
-              indexes[path] = {
-                liquid_drop: repo.to_liquid,
-              }
-            end
-
-            indexes[path]
-          end
+        def index_missing_message(path)
+          "Unable to load EXPRESS index for `#{path}`, " \
+            "please define it at `:lutaml-express-index:` or specify " \
+            "the full path."
         end
 
         def update_repo(options, repo)
-          # Unwrap repo if it's a cache
-          repo = repo.content if repo.is_a? Expressir::Model::Cache
+          repo = repo.content if repo.is_a?(Expressir::Model::Cache)
+          return repo unless repo.is_a?(Expressir::Model::Repository) ||
+            repo.is_a?(Expressir::Model::ExpFile)
 
-          # Process each schema
           repo.schemas.each do |schema|
             options["relative_path_prefix"] =
               relative_path_prefix(options, schema)
@@ -150,6 +65,25 @@ module Metanorma
           repo
         end
 
+        def template(lines)
+          ::Liquid::Template.parse(
+            lines.join("\n"),
+            environment: create_liquid_environment,
+          )
+        end
+
+        def reorder_schemas(repo_liquid, options)
+          return repo_liquid.schemas unless options["selected_schemas"]
+
+          options["selected_schemas"].filter_map do |schema_name|
+            repo_liquid.schemas.find do |schema|
+              schema.id == schema_name || schema.file_basename == schema_name
+            end
+          end
+        end
+
+        private
+
         def update_remarks(model, options)
           model.remarks = decorate_remarks(options, model.remarks)
           model.remark_items&.each do |ri|
@@ -157,127 +91,29 @@ module Metanorma
           end
 
           model.children.each do |child|
-            if child.respond_to?(:remarks) && child.respond_to?(:remark_items)
-              update_remarks(child, options)
-            end
+            next unless child.is_a?(Expressir::Model::ModelElement)
+
+            update_remarks(child, options)
           end
         end
 
         def relative_path_prefix(options, model)
-          return nil if options.nil? || options["document"].nil?
+          return if options.nil? || options["document"].nil?
 
           document = options["document"]
           file_path = File.dirname(model.file)
           docfile_directory = File.dirname(
             document.attributes["docfile"] || ".",
           )
-          document
-            .path_resolver
-            .system_path(file_path, docfile_directory)
+          document.path_resolver.system_path(file_path, docfile_directory)
         end
 
         def decorate_remarks(options, remarks)
           return [] unless remarks
 
           remarks.map do |remark|
-            ::Metanorma::Plugin::Lutaml::ExpressRemarksDecorator
-              .call(remark, options)
+            ExpressRemarksDecorator.call(remark, options)
           end
-        end
-
-        def read_config_yaml_file(document, file_path) # rubocop:disable Metrics/MethodLength
-          return {} if file_path.nil?
-
-          relative_file_path = Utils.relative_file_path(document, file_path)
-          config_yaml = YAML.safe_load(
-            File.read(relative_file_path, encoding: "UTF-8"),
-          )
-
-          return {} unless config_yaml["schemas"]
-
-          unless config_yaml["schemas"].is_a?(Hash)
-            raise StandardError.new(
-              "[lutaml_express_liquid] attribute `config_yaml` must point " \
-              "to a YAML file that has the `schemas` key containing a hash.",
-            )
-          end
-
-          { "selected_schemas" => config_yaml["schemas"].keys }
-        end
-
-        def render_liquid_template(document:, lines:, context_name:, # rubocop:disable Metrics/AbcSize,Metrics/MethodLength,Metrics/ParameterLists
-                      index_names:, options:, indexes:)
-          # Process options and configuration
-          options = process_options(document, options)
-
-          # Get all context items in one go
-          all_items = gather_context_liquid_items(
-            index_names: index_names, document: document, indexes: indexes,
-            options: options.merge("document" => document)
-          )
-
-          # Setup include paths for liquid templates
-          include_paths = [Utils.relative_file_path(document, "")]
-          options["include_path"]&.split(",")&.each do |path|
-            # resolve include_path relative to the document
-            include_paths.push(Utils.relative_file_path(document, path))
-          end
-
-          file_system = ::Metanorma::Plugin::Lutaml::Liquid::LocalFileSystem
-            .new(include_paths, ["%s.liquid", "_%s.liquid", "_%s.adoc"])
-
-          # Parse template once outside the loop
-          template = ::Liquid::Template
-            .parse(lines.join("\n"), environment: create_liquid_environment)
-          template.registers[:file_system] = file_system
-
-          # Render for each item
-          all_items.map do |item|
-            template.assigns[context_name] = item[:liquid_drop]
-            template.assigns["ordered_schemas"] = reorder_schemas(
-              item[:liquid_drop], options
-            )
-            template.assigns["schemas_order"] = options["selected_schemas"]
-            template.render
-          end.flatten
-        rescue StandardError => e
-          ::Metanorma::Util
-            .log("[LutamlPreprocessor] Failed to parse LutaML block: " \
-                 "#{e.message}", :error)
-          raise e
-        end
-
-        def reorder_schemas(repo_liquid, options)
-          return repo_liquid.schemas unless options["selected_schemas"]
-
-          ordered_schemas = []
-          options["selected_schemas"].each do |schema_name|
-            ordered_schema = repo_liquid.schemas.find do |schema|
-              schema.id == schema_name || schema.file_basename == schema_name
-            end
-            ordered_schemas.push(ordered_schema)
-          end
-
-          ordered_schemas
-        end
-
-        def process_options(document, options)
-          # Process config file if specified
-          if (config_yaml_path = options.delete("config_yaml"))
-            config = read_config_yaml_file(document, config_yaml_path)
-            if config["selected_schemas"]
-              options["selected_schemas"] =
-                config["selected_schemas"]
-            end
-          end
-          options
-        end
-
-        def parse_options(options_string)
-          options_string
-            .to_s
-            .scan(/,\s*([^=]+?)=(\s*[^,]+)/)
-            .to_h { |elem| elem.map(&:strip) }
         end
       end
     end
